AWS-AGENT-FACTORY/factory/iac/requirements.txt
aws-cdk-lib==2.204.0
constructs>=10.0.0,<11.0.0
pyyaml
aws-cdk.aws-lambda-go-alpha
